Feat/arbitrum (#1341)
* Added arbitrum secret keys

* Added arbitrum seed tokens

* Fixed test issue

* Added acceptFunOnArbitrumFilter

* Removed .only attribute from a test had remained by mistake

* Added token verification tests

* Updated validators to support arbitrum

* Fixed coingecko id of TrueUsd

* Changed a test description
