+++ .github/workflows/staging-pipeline.yml
@@ -100,6 +100,8 @@ jobs:
           OPTIMISTIC_SCAN_API_KEY: ${{ secrets.OPTIMISTIC_SCAN_API_KEY }}
           CELO_SCAN_API_KEY: ${{ secrets.CELO_SCAN_API_KEY }}
           CELO_ALFAJORES_SCAN_API_KEY: ${{ secrets.CELO_ALFAJORES_SCAN_API_KEY }}
+          ARBITRUM_SCAN_API_KEY: ${{ secrets.ARBITRUM_SCAN_API_KEY }}
+          ARBITRUM_SEPOLIA_SCAN_API_KEY: ${{ secrets.ARBITRUM_SEPOLIA_SCAN_API_KEY }}
           MORDOR_ETC_TESTNET: ${{ secrets.MORDOR_ETC_TESTNET }}
           ETC_NODE_HTTP_URL: ${{ secrets.ETC_NODE_HTTP_URL }}
           DROP_DATABASE: ${{ secrets.DROP_DATABASE_DURING_TEST_STAGING }}
