+++ config/example.env
@@ -36,6 +36,10 @@ CELO_SCAN_API_URL=https://api.celoscan.io/api
 CELO_SCAN_API_KEY=0000000000000000000000000000000000
 CELO_ALFAJORES_SCAN_API_URL=https://api-alfajores.celoscan.io/api
 CELO_ALFAJORES_SCAN_API_KEY=0000000000000000000000000000000000
+ARBITRUM_SCAN_API_URL=https://api.arbiscan.io/api
+ARBITRUM_SCAN_API_KEY=0000000000000000000000000000000000
+ARBITRUM_SEPOLIA_SCAN_API_URL=https://api-sepolia.arbiscan.io/api
+ARBITRUM_SEPOLIA_SCAN_API_KEY=0000000000000000000000000000000000
 GNOSISSCAN_API_URL=https://api.gnosisscan.io/api
 ETHERSCAN_API_KEY=0000000000000000000000000000000000
 GNOSISSCAN_API_KEY=0000000000000000000000000000000000
@@ -223,6 +227,11 @@ MORDOR_ETC_TESTNET_SCAN_API_URL=https://etc-mordor.blockscout.com/api/v1
 # https://chainlist.org/chain/63
 MORDOR_ETC_TESTNET_NODE_HTTP_URL=https://rpc.mordor.etccooperative.org
 
+# ARBITRUM MAINNET
+ARBITRUM_MAINNET_NODE_HTTP_URL=https://arb1.arbitrum.io/rpc
+
+# ARBITRUM SEPOLIA
+ARBITRUM_SEPOLIA_NODE_HTTP_URL=https://sepolia-rollup.arbitrum.io/rpc
 
 # This is the address behind donation.eth
 MATCHING_FUND_DONATIONS_FROM_ADDRESS=0x6e8873085530406995170Da467010565968C7C62
