+++ config/test.env
@@ -44,6 +44,10 @@ POLYGON_SCAN_API_URL=https://api.polygonscan.com/api
 OPTIMISTIC_SCAN_API_URL=https://api-optimistic.etherscan.io/api
 CELO_SCAN_API_URL=https://api.celoscan.io/api
 CELO_ALFAJORES_SCAN_API_URL=https://api-alfajores.celoscan.io/api
+ARBITRUM_SCAN_API_URL=https://api.arbiscan.io/api
+ARBITRUM_SCAN_API_KEY=0000000000000000000000000000000000
+ARBITRUM_SEPOLIA_SCAN_API_URL=https://api-sepolia.arbiscan.io/api
+ARBITRUM_SEPOLIA_SCAN_API_KEY=0000000000000000000000000000000000
 GNOSISSCAN_API_URL=https://api.gnosisscan.io/api
 ETHERSCAN_API_KEY=00000000000000000000000000000000
 GNOSISSCAN_API_KEY=0000000000000000000000000000000000
