+++ migration/data/seedTokens.ts
@@ -13,6 +13,7 @@ interface ITokenData {
   networkId: number;
   chainType?: ChainType;
   coingeckoId?: string;
+  isStableCoin?: boolean;
 }
 const seedTokens: ITokenData[] = [
   // Mainnet tokens
@@ -1375,6 +1376,238 @@ const seedTokens: ITokenData[] = [
     networkId: NETWORK_IDS.SOLANA_TESTNET,
     chainType: ChainType.SOLANA,
     coingeckoId: COINGECKO_TOKEN_IDS.SOLANA,
+  },
+
+  // ARBITRUM Sepolia
+  {
+    name: 'Arbitrum Sepolia native token',
+    symbol: 'ETH',
+    address: '0x0000000000000000000000000000000000000000',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_SEPOLIA,
+    coingeckoId: 'ethereum',
+  },
+  {
+    name: 'Chromatic test Eth',
+    symbol: 'cETH',
+    address: '0x93252009E644138b906aE1a28792229E577239B9',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_SEPOLIA,
+    coingeckoId: 'weth',
+  },
+  {
+    name: 'Arbitrum ETH',
+    symbol: 'ETH',
+    address: '0x0000000000000000000000000000000000000000',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'ethereum',
+  },
+  {
+    name: 'usdt',
+    symbol: 'USDT',
+    address: '0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9',
+    decimals: 6,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'tether',
+    // isStableCoin: true,
+  },
+  {
+    name: 'USDC',
+    symbol: 'USDC',
+    address: '0xaf88d065e77c8cC2239327C5EDb3A432268e5831',
+    decimals: 6,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'usd-coin',
+    // isStableCoin: true,
+  },
+  {
+    name: 'Chainlink',
+    symbol: 'LINK',
+    address: '0xf97f4df75117a78c1A5a0DBb814Af92458539FB4',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'chainlink',
+  },
+  {
+    name: 'Dai',
+    symbol: 'DAI',
+    address: '0xDA10009cBd5D07dd0CeCc66161FC93D7c9000da1',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'dai',
+    // isStableCoin: true,
+  },
+  {
+    name: 'Uniswap',
+    symbol: 'UNI',
+    address: '0xFa7F8980b0f1E64A2062791cc3b0871572f1F7f0',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'uniswap',
+  },
+  {
+    name: 'Lido',
+    symbol: 'LDO',
+    address: '0x13Ad51ed4F1B7e9Dc168d8a00cB3f4dDD85EfA60',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'lido-dao',
+  },
+  {
+    name: 'Arbitrum',
+    symbol: 'ARB',
+    address: '0x912CE59144191C1204E64559FE8253a0e49E6548',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'arbitrum',
+  },
+  {
+    name: 'Brigded ISDC',
+    symbol: 'USDC',
+    address: '0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8',
+    decimals: 6,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'usd-coin',
+    // isStableCoin: true,
+  },
+  {
+    name: 'TrueUSD',
+    symbol: 'TUSD',
+    address: '0x4D15a3A2286D883AF0AA1B3f21367843FAc63E07',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'true-usd',
+    // isStableCoin: true,
+  },
+  {
+    name: 'The Graph',
+    symbol: 'GRT',
+    address: '0x9623063377AD1B27544C965cCd7342f7EA7e88C7',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'the-graph',
+  },
+  {
+    name: 'Frax Share',
+    symbol: 'FXS',
+    address: '0x9d2f299715d94d8a7e6f5eaa8e654e8c74a988a7',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'frax-share',
+  },
+  {
+    name: 'USDD',
+    symbol: 'USDD',
+    address: '0x680447595e8b7b3Aa1B43beB9f6098C79ac2Ab3f',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'usdd',
+    // isStableCoin: true,
+  },
+  {
+    name: 'WOO',
+    symbol: 'WOO',
+    address: '0xcafcd85d8ca7ad1e1c6f82f651fa15e33aefd07b',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'woo-network',
+  },
+  {
+    name: 'Gnosis',
+    symbol: 'GNO',
+    address: '0xa0b862f60edef4452f25b4160f177db44deb6cf1',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'gnosis',
+  },
+  {
+    name: 'Curve DAO Token',
+    symbol: 'CRV',
+    address: '0x11cDb42B0EB46D95f990BeDD4695A6e3fA034978',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'curve-dao-token',
+  },
+  {
+    name: 'Compound',
+    symbol: 'COMP',
+    address: '0x354A6dA3fcde098F8389cad84b0182725c6C91dE',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'compound-governance-token',
+  },
+  {
+    name: 'GMX',
+    symbol: 'GMX',
+    address: '0xfc5a1a6eb076a2c7ad06ed22c90d7e710e35ad0a',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'gmx',
+  },
+  {
+    name: 'Loopring',
+    symbol: 'LRC',
+    address: '0x46d0cE7de6247b0A95f67b43B589b4041BaE7fbE',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'loopring',
+  },
+  {
+    name: 'Treasure',
+    symbol: 'MAGIC',
+    address: '0x539bdE0d7Dbd336b79148AA742883198BBF60342',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'treasure',
+  },
+  {
+    name: 'SushiSwap',
+    symbol: 'SUSHI',
+    address: '0xd4d42f0b6def4ce0383636770ef773390d85c61a',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'sushi',
+  },
+  {
+    name: 'yearn.finance',
+    symbol: 'YFI',
+    address: '0x82e3A8F066a6989666b031d916c43672085b1582',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'yearn-finance',
+  },
+  {
+    name: 'Xai',
+    symbol: 'XAI',
+    address: '0x4Cb9a7AE498CEDcBb5EAe9f25736aE7d428C9D66',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'xai',
+  },
+  {
+    name: 'Livepeer',
+    symbol: 'LPT',
+    address: '0x289ba1701C2F088cf0faf8B3705246331cB8A839',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'livepeer',
+  },
+  {
+    name: 'Wrapped Bitcoin',
+    symbol: 'WBTC',
+    address: '0x2f2a2543B76A4166549F7aaB2e75Bef0aefC5B0f',
+    decimals: 8,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'wrapped-bitcoin',
+  },
+  {
+    name: 'Cartesi',
+    symbol: 'CTSI',
+    address: '0x319f865b287fCC10b30d8cE6144e8b6D1b476999',
+    decimals: 18,
+    networkId: NETWORK_IDS.ARBITRUM_MAINNET,
+    coingeckoId: 'cartesi',
   },
 ];
