+++ src/server/adminJs/tabs/donationTab.ts
@@ -604,6 +604,8 @@ export const donationTab = {
           { value: NETWORK_IDS.POLYGON, label: 'Polygon' },
           { value: NETWORK_IDS.CELO, label: 'Celo' },
           { value: NETWORK_IDS.CELO_ALFAJORES, label: 'Alfajores' },
+          { value: NETWORK_IDS.ARBITRUM_MAINNET, label: 'Arbitrum' },
+          { value: NETWORK_IDS.ARBITRUM_SEPOLIA, label: 'Arbitrum Sepolia' },
         ],
         isVisible: {
           list: true,
