+++ src/server/adminJs/tabs/qfRoundTab.ts
@@ -133,6 +133,8 @@ export const qfRoundTab = {
             value: NETWORK_IDS.CELO_ALFAJORES,
             label: 'ALFAJORES (Test CELO)',
           },
+          { value: NETWORK_IDS.ARBITRUM_MAINNET, label: 'ARBITRUM MAINNET' },
+          { value: NETWORK_IDS.ARBITRUM_SEPOLIA, label: 'ARBITRUM SEPOLIA' },
           { value: NETWORK_IDS.XDAI, label: 'XDAI' },
           { value: NETWORK_IDS.BSC, label: 'BSC' },
         ],
