+++ src/server/adminJs/tabs/tokenTab.ts
@@ -190,6 +190,8 @@ export const generateTokenTab = async () => {
               value: NETWORK_IDS.CELO_ALFAJORES,
               label: 'ALFAJORES (Test CELO)',
             },
+            { value: NETWORK_IDS.ARBITRUM_MAINNET, label: 'ARBITRUM MAINNET' },
+            { value: NETWORK_IDS.ARBITRUM_SEPOLIA, label: 'ARBITRUM SEPOLIA' },
             { value: NETWORK_IDS.XDAI, label: 'XDAI' },
             { value: NETWORK_IDS.BSC, label: 'BSC' },
             { value: NETWORK_IDS.ETC, label: 'Ethereum Classic' },
