+++ src/utils/validators/graphqlQueryValidators.ts
@@ -220,6 +220,8 @@ const managingFundsValidator = Joi.object({
         NETWORK_IDS.POLYGON,
         NETWORK_IDS.CELO,
         NETWORK_IDS.CELO_ALFAJORES,
+        NETWORK_IDS.ARBITRUM_MAINNET,
+        NETWORK_IDS.ARBITRUM_SEPOLIA,
         NETWORK_IDS.OPTIMISTIC,
         NETWORK_IDS.OPTIMISM_GOERLI,
         NETWORK_IDS.XDAI,
