+++ src/utils/validators/projectValidator.test.ts
@@ -73,14 +73,22 @@ function isWalletAddressSmartContractTestCases() {
     assert.isTrue(isSmartContract);
   });
   it('should return true for smart contract address in celo', async () => {
-    // GIV address https://polygonscan.com/address/0x0d500b1d8e8ef31e21c99d1db9a6444d3adf1270
     const walletAddress = '0x67316300f17f063085Ca8bCa4bd3f7a5a3C66275';
     const isSmartContract = await isWalletAddressSmartContract(walletAddress);
     assert.isTrue(isSmartContract);
   });
   it('should return true for smart contract address in celo alfajores', async () => {
-    // GIV address https://polygonscan.com/address/0x0d500b1d8e8ef31e21c99d1db9a6444d3adf1270
     const walletAddress = '0x17bc3304F94c85618c46d0888aA937148007bD3C';
+    const isSmartContract = await isWalletAddressSmartContract(walletAddress);
+    assert.isTrue(isSmartContract);
+  });
+  it('should return true for smart contract address in arbitrum mainnet', async () => {
+    const walletAddress = '0x1231DEB6f5749EF6cE6943a275A1D3E7486F4EaE';
+    const isSmartContract = await isWalletAddressSmartContract(walletAddress);
+    assert.isTrue(isSmartContract);
+  });
+  it('should return true for smart contract address in arbitrum sepolia', async () => {
+    const walletAddress = '0x6b7860b66c0124e8d8c079b279c126ce58c442a2';
     const isSmartContract = await isWalletAddressSmartContract(walletAddress);
     assert.isTrue(isSmartContract);
   });
